refactor(site): moves theme components
- also reserve space for lazy theme icon before loading
